Introduce first of the fruitpile operations
The basic operations of initialization of a repo, adding a file to
it and listing (some of) the contents of the repo are now
available.  This enables the whole thing to actually start being
vaguely usable.  However, there's not a huge amount of sanity
checking so it's not suitable to actually store anything in yet.