Add in basic schema for permissions
Add in basic schema support for permissioning.  A user is created
in the user table and linked to one of a number of permissions
through the user permissions table (user_perms).  Permissions and
users are both required to have unique names.  The user permissions
table also requires that the combination of user id and permission
id are unique.

The full list of permissions are accessed through an AllPermissions
object as if they are a kind of enumerated type.  For the purposes
of Fruitpile, permission names are upper case and made only from
characters and underscores.  They are also limited to 20 characters
long.