Reworked permissions to be a Python constant
Python now drives the settings of permission ids.  This allows for
a simpler process when upgrading since the permissions can be
maintained with static ids over upgrades.  This avoids the problem
of the database and the application not having the same
understanding of the ids.  The two can of course still get in a
mess but this makes the code somewhat easier to reasona about (the
database table just reflects the application and is used to prevent
the values in the user permissions table becoming corrupt - you can
only change the value of a permission to an already declared value.
Note also that the name has changed to Capability to reflect that
this represents the ability to do something, while the user is
granted a Permission to exercise that capability.