Add in state transition details
Add in database tables to manage state transitions.  Will do no
more on this but for now the transitions table simply tracks the
start and end state id and if a transition function/check is to be
called then this can be added into the transition record (arc in
the state graph).