Added tests for simple state machine
Added simple state machine to support transition of states of the
files under control.  This state machine allows the states to be
transited under the control of both the permission manager and also
a callback function.  In principle the callback function can do
anything and carries with it user id, the permission manager,
the old and new states and an object which can carry whatever else
is needed (probably most likely in Fruitpile to be a BinFile
of FileSet object to transition).

Currently there is no integration with the database data at this
point.