Added transition state machine to static data
Relocated static data initialization to a seperate module to avoid
the Fruitpile.init method becoming overly long and complex.  Added
both states and transitions for the default set that is planned for
Fruitpile first release.

Add functionality to allow Fruitpile to pull in the state machine
and transitions as well as capabilities to ensure that all those
operations can be controlled through the permissining engine.

Confirmed that the basic configuration of the state machine appears
as expected.