Add ability to transit a file's state
Hook in the file state machine into the file transit operation
and validate that a file's state can be transitioned between
different states.  In this code currently there is no support
for performing additionaly checks alongside the basic legitimacy
ones for state transitioning (e.g. for checking for auxilliary
files).