Correct interpretation of transfn present or not
On Travis.CI all python versions were passing except 3.3 which
showed up a wierd error to do with an item not having a bound
session.

It turns out that the ORM from Sqlalchemy was correct on 3.3 and
allowing something to come through broken on other platforms.  The
relationship between the transition and the transition function
is one of a foreign key allowing nulls.  Because the relationship
allows nulls the transfn can potentially not be present or being
present as None or some other variety of absence.  The correct
check should be on the transfn_id which is the actual linking
column which is the one that can be null or not null accordingly.