Add fp_tool for simple repo operation
Prior to the REST API being available this change provides the basis
of a command line too that will allow a user to operate a Fruitpile
store.  The command line tool is a bit rubbish at this point but
it is at least usable and will probably meet the purpose originally
intended for Fruitpile while other work on Fruitpile will be done
incrementally.