Use Jinja2 templating to output results of list ops
For listing operations use the Jinja2 templating system to prepare
the output text as Jinja will be there due to Flask dependency and
it's easier than writing all the formatting strings.

This does not currently work with Python 3.2 (though may with 3.3+)
as this is the version which Jinja2 has passed its unit tests on.