Move version and revision to fileset from binfile
Move the version and revision values to the filesets table from
the binfiles table.  There are two good reasons for this.  The
first is that a fileset is meant to encompass the notion of the
'output from a given build'.  This means that the idea that the
outputs are somehow different and from different builds doesn't
really make much sense.  The second is that there would be a huge
amount of duplicated data, and in the case of auxilliary files
they do not have a version or revision themselves - they simply
pertain to a version/revision combination.  (Also is should also
be observed that the version should be able to be directly
determined from the revision as revision is intended to be the
revision in the source code tree.  However, version is stored in
the repository database simply to make it easier to find.)

It is also possible that when properties are introduced the version
and revision may just become a property and therefore not
special in anyway.