Remove support for Python 3.2
Without the facility to use u"this is a unicode string" in Python 3.2
writing code that works on both 2.7 and 3.2 requires a lot more
effort in terms of supports unicode.  The simpler answer is to
simply remove 3.2 support in travis and only support 3.3 upwards.
This also means that Jinja appears to work well enough on 3.3 to
allow the tests to pass.