Provide tests to cover fp_tool
Unit test fp_tool.  In the process move most of the fp_tool code
into the fruitpile module.  This allows the unit test code to
import the body of all the functionality so tests can be carried
out on the main functions as well as on example command lines.
This results in the real fp_tool being a tiny shell of a script
that simply invokes the main function of fruitpile/fp_tool.py.

Removed support for Python 3.2 allows unit tests to run on
Python 3.3 and upwards (apparently).  And these tests run across
pypy, 2.7 and 3.3.

Modify the output a little to make it somewhat more straigtforward
to read and relocate the revision and version informatio to the
fileset list rather than the file list.