Fix py3 tests
Due to a mix-up in python virtual environments the python 3 venv
was actually running the py.test executable which was in turn
bringing in python 2.7 and thus reporting that is was successful
but was not in fact running against a python 3 interpreter.  The
virtual env is now fixed and the same error as in travis has
shown up.  This change corrects that problem.