fp_tool now supports state transitions
It is now possible to transition states through the fp_tool
command line tool.  The state machine needs to be fed the existing
state to know how to transition the state (and this came up through
testing).  By default the state machine starts off "untested" and
the state wasn't being fed back into the state machine once the
first transition was made.