Add transition function data to schema
Added table to store the data that is to be used by the the
transition function when looking at transitioning from one state
to another.  This data is handed into the transition function
alongside the objcet being transitioned and the desired state
along with a few other parameters.  This transition function allows
additional qualifications to be checked when transitioning states.