Auxilliary file check now on state transition
It is now possible to confirm before transitioning to a new state
that there is an additional set of criteria fulfilled.  Currently
the only check conducted is to ensure that there is an
appropriately named auxilliary file.  The check only ensures that
the name ends with the requisite character string.  It is possible
to ensure other things but currently this is the only function
provided.