Beginning the REST API
Added flask_restful as a dependency and building the REST API
around that functionality.  The REST API is being reported as v1
and lies under the v1 tree.  In order to support the flask
development, relaxing the restriction on only a single accessor
to the repository (which was going to be dropped anyhow).  Remove
the related tests and rework the exclusive access test to one that
proves duplicate access.

Provided a simple endpoint that lists all files which currently
works correctly with an empty respository.  However, there is no
mapping of binary files back to anything sensible.