Added a very basic authentication scheme
Added support for authenticating a user to the Fruitpile REST API.
This is simply to have the authentication process in place so that
the remainder of the API can be built around the notion of the
login capability.  For the primary use case this is largely to just
get the uid of the originator more than anything else.