Add tags tests
Add tests to cover tagging a fileset.  Confirm that tags can be
added to a fileset, that the tag is created automatically if it
doesn't exist, is reused if it does exist and that it is shared
if multiple filesets use the same tag.

From the outside world tags are just strings.