Add tags and properties to artifacts
This change adds the ability to add tags and properties to artifacts
as well as filesets.  The tags can overlap with those on filesets
(as the tag is only stored once).  Properties however, are most
likely to be distinct by virtue of the fact that they are a key
value pair.